MINOR: Change insert prototype and unique become implicit
Non unicity were managed throught an array of interface{}, where the
first node has first place and other node have next places. This is
now on the caller side, so the "unique" no longer exists, each key is
always unique.

The prototype of insert function change: the insert function returns
two values: node and inserted bool. The node is the concerned node,
the bool indicate if an insertion was done or match existing node.