BUG/MEDIUM: The exact match is not rightly handled
If the inserted radix is the same than compared radix, but the inserted
mask is shorter than the compared mask, the exact match is match, but its
wrong.

This bad decision could triger a segfault is the compared node has a radix
number of byte greater than the inserted node.